Move fetchSavedQueryConfig out into its own util
