+++ src/routes/SavedQueriesRoute.js
@@ -1,10 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { useStripes } from '@folio/stripes/core';
 import { LoadingPane } from '@folio/stripes/components';
-import stripesFetch from '../util/stripesFetch';
 import BigError from '../components/BigError';
+import fetchSavedQueryConfig from '../util/fetchSavedQueryConfig';
 import SavedQueries from '../components/SavedQueries';
-
 
 function SavedQueriesRoute() {
   const stripes = useStripes();
@@ -12,18 +11,7 @@ function SavedQueriesRoute() {
   const [error, setError] = useState();
 
   useEffect(() => {
-    async function fetchData() {
-      const res = await stripesFetch(stripes, '/ldp/config/sqconfig');
-      if (res.ok) {
-        const json = await res.json();
-        const data = JSON.parse(json.value);
-        setConfig(data);
-      } else {
-        const content = await res.text();
-        setError(`${res.statusText}: ${content}`);
-      }
-    }
-    fetchData();
+    fetchSavedQueryConfig(stripes, setConfig, setError);
   }, [stripes]);
 
   if (error) return <BigError message={error} />;
